Add AudioProcessSupplier class for managing audio process discovery and selection. Update CoreAudioTapEngineExamples to utilize the new supplier for process management and implement auto-stop functionality. Adjust CaptionView to start system audio capture on view appearance.
